requirements.txt
flask==3.1.0
openai==1.68.2
python-dotenv==1.0.1
flask-cors==5.0.1
